currently doing role based authentication. need to test

+++ backend/routes/patients.js
@@ -4,6 +4,9 @@ const requireAuth = require("../middleware/requireAuth");
 
 router.use(requireAuth);
 router.route("/add").post((req, res) => {
+
+  if((req.user.role !== 'admin')||(req.user.role !== 'staff')) return res.status(403).send("You are not allowed to make these changes");
+
   const fullName = req.body.fullName;
   const gender = req.body.gender;
   const dateOfBirth = Date(req.body.dateOfBirth);
@@ -55,21 +58,28 @@ router.route("/add").post((req, res) => {
 })
 
 router.route("/get/:nicNumber").get(async (req, res) => {
+
+  if((req.user.role !== 'admin')||(req.user.role !== 'staff')) return res.status(403).send("You are not allowed to make these changes");
+
+
   const nicNumber = req.params.nicNumber;
   const user = await patient.findOne({ nicNumber })
-  .then((user) => {
-    if (!user) {
-      return res.status(404).json({ error: "Patient not found" })
-    }
-    res.status(200).json({ status: "Patient fetched", user });
-  })
-  .catch((err) => {
-    console.log(err.message);
-    res.status(500).json({ status: "Error with fetching patient", error: err.message });
-  });
+    .then((user) => {
+      if (!user) {
+        return res.status(404).json({ error: "Patient not found" })
+      }
+      res.status(200).json({ status: "Patient fetched", user });
+    })
+    .catch((err) => {
+      console.log(err.message);
+      res.status(500).json({ status: "Error with fetching patient", error: err.message });
+    });
 });
 
 router.route("/update/:nicNumber").put(async (req, res) => {
+  
+  if((req.user.role !== 'admin')||(req.user.role !== 'staff')) return res.status(403).send("You are not allowed to make these changes");
+
   const nicNumber = req.params.nicNumber;
   const {
     fullName,
@@ -125,7 +135,7 @@ router.route("/update/:nicNumber").put(async (req, res) => {
     if (updatedPatient) {
       res.status(200).send({ status: "Patient updated", data: updatedPatient });
     } else {
-      res.status(404).send({ status: "Patient not found"});
+      res.status(404).send({ status: "Patient not found" });
     }
   } catch (err) {
     console.log(err);
@@ -134,6 +144,9 @@ router.route("/update/:nicNumber").put(async (req, res) => {
 });
 
 router.route("/delete/:nicNumber").delete(async (req, res) => {
+  
+  if((req.user.role !== 'admin')||(req.user.role !== 'staff')) return res.status(403).send("You are not allowed to make these changes");
+
   const nicNumber = req.params.nicNumber;
   try {
     const patient = await patient.findOneAndDelete({ nicNumber });
@@ -148,6 +161,10 @@ router.route("/delete/:nicNumber").delete(async (req, res) => {
 });
 
 router.route("/").get((req, res) => {
+  
+  if((req.user.role !== 'admin')||(req.user.role !== 'staff')) return res.status(403).send("You are not allowed to make these changes");
+
+
   patient.find()
     .then((patient) => {
       res.json(patient);
